changed all forms to wtforms and enabled users to change personal data and delete acount

+++ app.py
@@ -1,11 +1,13 @@
-from flask import Flask, render_template, request, redirect, url_for, flash, escape, sessions, abort
+from flask import Flask, Blueprint, render_template, request, redirect, url_for, flash, escape, sessions, abort
+from jinja2 import TemplateNotFound
 from database import db, app
-from models import User
-from werkzeug.security import generate_password_hash, check_password_hash
+from models import User, authenticate
+from werkzeug.security import generate_password_hash
 from flask_login import login_manager, LoginManager, AnonymousUserMixin, login_url, login_user, logout_user, login_required, current_user
 from flask_wtf import FlaskForm, Form
 from flask_wtf.csrf import CsrfProtect
 from wtforms import StringField, PasswordField, SubmitField, validators
+from forms import LoginForm, SignupForm, EditForm, DeleteForm
 
 
 login_manager = LoginManager()
@@ -38,26 +40,22 @@ def profile():
     return render_template('users/profile.html', user_full_name=user_full_name)
 
 
-@app.route("/users/new")
-def new():
-    # if current_user
-    return render_template('users/new.html')
-
-
-@app.route("/users", methods=["POST"])
+@app.route("/users/new", methods=['GET', 'POST'])
 def create():
-    full_name = request.form.get('full_name')
-    email = request.form.get('email')
-    username = request.form.get('username')
-    password = request.form.get('password')
-    hashed_password = generate_password_hash(password)
-
-    new_user = User(full_name, email, username, hashed_password)
-    db.session.add(new_user)
-    db.session.commit()
-    login_user(new_user)
-
-    return redirect(url_for('profile', user_full_name=new_user.full_name))
+    if current_user.is_authenticated:
+        flash('You must be logged out to sign up!', 'warning')
+        return redirect(url_for('home'))
+    else:
+        form = SignupForm()
+        if form.validate_on_submit():
+            hashed_password = generate_password_hash(form.password.data)
+            new_user = User(form.full_name.data, form.email.data,
+                            form.username.data, hashed_password)
+            db.session.add(new_user)
+            db.session.commit()
+            login_user(new_user)
+            return redirect(url_for('profile'))
+        return render_template('users/new.html', form=form)
 
 
 @app.route("/users", methods=["GET"])
@@ -71,24 +69,6 @@ def index():
 def show(id):
     user = User.query.get(id)
     return render_template('users/show.html', user=user)
-
-
-def authenticate(username, password):
-    try:
-        user = User.query.filter_by(username=username).all()[0]
-        password_check = check_password_hash(user.password, password)
-        if password_check:
-            return user
-        else:
-            flash('Wrong Username/Password!', 'Warning')
-    except IndexError:
-        flash('Wrong Username/Password!', 'Warning')
-
-
-class LoginForm(FlaskForm):
-    username = StringField('Username', [validators.InputRequired()])
-    password = PasswordField('Password', [validators.InputRequired()])
-    submit = SubmitField('Submit')
 
 
 @app.route('/login', methods=['GET', 'POST'])
@@ -118,32 +98,46 @@ def logout():
     return redirect(url_for('home'))
 
 
-@app.route("/users/<id>/edit", methods=["GET"])
-@login_required
-def edit(id):
-    user = User.query.get(id)
-    return render_template('users/edit.html', user=user)
+# @app.route("/users/<id>/edit", methods=["GET"])
+# @login_required
+# def edit(id):
+#     user = User.query.get(id)
+#     return render_template('users/edit.html', user=user)
 
 
-@app.route("/users/<id>", methods=["POST"])
+@app.route("/users/<id>/edit", methods=["GET", "POST"])
 @login_required
 def update_or_destroy(id):
+    if request.args.get('_method') == 'PUT':
+        form = EditForm()
+        return render_template('users/edit.html', User=User, id=id, form=form)
     if request.form.get('_method') == 'PUT':
-        editted_user = User.query.get(id)
-        editted_user.full_name = request.form.get('full_name')
-        editted_user.email = request.form.get('email')
-        editted_user.username = request.form.get('username')
-        editted_user.password = generate_password_hash(
-            request.form.get('password'))
-        db.session.add(editted_user)
-        db.session.commit()
-        return redirect(url_for('index'))
-
-    elif request.form.get('_method') == 'DELETE':
-        user = User.query.get(id)
-        db.session.delete(user)
-        db.session.commit()
-        return redirect(url_for('index'))
+        form = EditForm()
+        if form.validate_on_submit():
+            editted_user = User.query.get(id)
+            editted_user.full_name = form.full_name.data
+            editted_user.email = form.email.data
+            editted_user.username = form.username.data
+            editted_user.password = generate_password_hash(
+                form.password.data)
+            db.session.add(editted_user)
+            db.session.commit()
+            flash('User details updated successfully.')
+            next = request.args.get('next')
+            return redirect(next or url_for('profile'))
+        return render_template('users/edit.html', User=User, id=id, form=form)
+    if request.args.get('_method') == 'DELETE':
+        form = DeleteForm()
+        return render_template('users/delete.html', id=id, form=form)
+    if request.form.get('_method') == 'DELETE':
+        form = DeleteForm()
+        if form.validate_on_submit():
+            user = User.query.get(id)
+            logout_user()
+            db.session.delete(user)
+            db.session.commit()
+            return redirect(url_for('home'))
+        return render_template('users/delete.html', id=id, form=form)
 
 
 if __name__ == '__main__':
